+++ modules/core/entity/modules/fields/farm_entity_fields.post_update.php
@@ -58,3 +58,21 @@ function farm_entity_fields_post_update_install_term_file_fields(&$sandbox) {
     }
   }
 }
+
+/**
+ * Install taxonomy term ontology URI field.
+ */
+function farm_entity_fields_post_update_add_term_ontology_uri(&$sandbox) {
+  $field_info = [
+    'type' => 'uri',
+    'label' => t('Ontology URI'),
+    'description' => t('Link this term to one or more external ontology item URIs.'),
+    'multiple' => TRUE,
+    'weight' => [
+      'form' => 80,
+      'view' => 80,
+    ],
+  ];
+  $field_definition = \Drupal::service('farm_field.factory')->baseFieldDefinition($field_info);
+  \Drupal::entityDefinitionUpdateManager()->installFieldStorageDefinition('ontology_uri', 'taxonomy_term', 'farm_entity_fields', $field_definition);
+}